+++ src/test/java/lotto/domain/LottoBundleTest.java
@@ -1,8 +1,6 @@
-package lotto;
+package lotto.domain;
 
 import org.junit.jupiter.api.Test;
-
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
