+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,10 +1,8 @@
-package lotto;
+package lotto.domain;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
